requirements.txt
google-api-python-client==2.108.0
youtube_transcript_api==0.6.1
python-dotenv==1.0.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
